+++ b3j0f/conf/__init__.py
@@ -30,6 +30,6 @@
 __all__ = ['__version_info__', '__version__']
 
 #: project version info
-__version_info__ = 0, 1, 6, "beta", 0
+__version_info__ = 0, 1, 7, "beta", 0
 #: project version
 __version__ = ".".join(str(v) for v in __version_info__[:3])