+++ b3j0f/conf/configurable/test/core.py
@@ -196,6 +196,62 @@ class ConfigurableTest(TestCase):
         self.configurable.configure(conf=conf)
         self.assertEqual(self.configurable.log_lvl, 'DEBUG')
 
+    def test_configure_to_reconfigure_param(self):
+        """Test to reconfigure an object with to_configure parameter.
+        """
+
+        self.assertTrue(self.configurable.auto_conf)
+
+        class ToConfigure(object):
+            """Class to configure.
+            """
+            def __init__(self):
+                super(ToConfigure, self).__init__()
+                self.test = None
+
+        to_configure = ToConfigure()
+
+        param = 'test'
+
+        conf = Configuration(
+            Category(
+                'TEST',
+                Parameter(param, value=True)
+            )
+        )
+
+        self.configurable.configure(conf=conf, to_configure=to_configure)
+        self.assertTrue(to_configure.test)
+
+    def test_configure_without_inheritance(self):
+        """Test to configure an object without inheritance.
+        """
+
+        class ToConfigure(object):
+            """Class to configure.
+            """
+            def __init__(self):
+
+                super(ToConfigure, self).__init__()
+
+                self.test = None
+
+        to_configure = ToConfigure()
+
+        configurable = Configurable(to_configure=to_configure)
+
+        param = 'test'
+
+        conf = Configuration(
+            Category(
+                'TEST',
+                Parameter(param, value=True)
+            )
+        )
+
+        configurable.configure(conf=conf, to_configure=to_configure)
+        self.assertTrue(to_configure.test)
+
     def test_parser_inheritance(self):
 
         class _Configurable(Configurable):
@@ -215,7 +271,8 @@ class ConfigurableTest(TestCase):
 
         self.assertEqual(
             len(configurable.conf) + 1,
-            len(_configurable.conf))
+            len(_configurable.conf)
+        )
 
 if __name__ == '__main__':
     main()